refactored form component to remove cycle dependency
