+++ client/app/widgets/QuickCreate/index.jsx
@@ -5,7 +5,6 @@ import { Input } from '../../components/Input';
 import type { Checkbox } from '../../components/Input/utils';
 import { Utils } from '../../utils';
 import css from './QuickCreate.scss';
-// eslint-disable-next-line import/no-cycle
 import { DynamicForm } from '../../components/Form/DynamicForm';
 
 // value - e.g. category.id
